Clippy fixes

+++ programs/incept/src/instructions/add_liquidity_to_single_pool_comet.rs
@@ -78,7 +78,7 @@ pub fn execute(
     let mut comet = ctx.accounts.single_pool_comet.load_mut()?;
 
     let pool_index = comet.positions[position_index as usize].pool_index as usize;
-    let empty_pool = token_data.pools[pool_index as usize].is_empty();
+    let empty_pool = token_data.pools[pool_index].is_empty();
 
     let usdi_liquidity_value = Decimal::new(usdi_amount.try_into().unwrap(), DEVNET_TOKEN_SCALE);
     let iasset_amm_value = Decimal::new(
@@ -220,7 +220,7 @@ pub fn execute(
         DEVNET_TOKEN_SCALE,
     );
 
-    let pool = token_data.pools[pool_index as usize];
+    let pool = token_data.pools[pool_index];
 
     if !empty_pool {
         return_error_if_false!(
